requirements.txt
crewai[tools]>=0.102.0,<1.0.0
pydantic
python-dotenv
hatchling
requests
arxiv 
tweepy>=4.14.0 
langgraph>=0.0.35
qdrant-client>=1.8.0
sentence-transformers
llama-index>=0.10.0
ollama
langchain
langchain-community
langchain-huggingface 
langchain-qdrant
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6 